update
modifs dans core/index.js
