+++ GenForm-NPM/Core/package.json
@@ -1,7 +1,8 @@
 {
   "name": "@genform/core",
-  "version": "0.0.4",
+  "version": "0.0.10",
   "description": "A form generator using JS based on a JSON object",
+  "private" : false,
   "main": "index.js",
   "scripts": {
     "minify": "uglifyjs --compress --mangle -o ./index.min.js -- ./index.js",
